refactor: move main logic to the /shared/directives/lcp-context. and leave in /cms-structure only the cms flavoured logic
