+++ projects/storefrontlib/cms-components/content/banner/banner.module.ts
@@ -12,9 +12,9 @@ import {
   FeaturesConfigModule,
   provideDefaultConfig,
 } from '@spartacus/core';
-import { LcpContextDirectiveModule } from '../../../cms-structure/lcp-context/lcp-context-directive.module';
 import { GenericLinkModule } from '../../../shared/components/generic-link/generic-link.module';
 import { MediaModule } from '../../../shared/components/media/media.module';
+import { LcpContextDirectiveModule } from '../../../shared/directives/lcp-context/lcp-context-directive.module';
 import { BannerComponent } from './banner.component';
 
 @NgModule({
