+++ projects/storefrontlib/cms-structure/cms-lcp-context/index.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2025 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './cms-lcp.service';
+export * from './provide-lcp-context-for-cms-directive.module';
+export * from './provide-lcp-context-for-cms.directive';
