+++ projects/storefrontlib/cms-structure/index.ts
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export * from './cms-lcp-context/index';
 export * from './guards/index';
-export * from './lcp-context/index';
 export * from './outlet/index';
 export * from './page/index';
 export * from './pwa/index';
