+++ projects/storefrontlib/cms-structure/services/index.ts
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './cms-guards.service';
-export * from './cms-i18n.service';
 export * from './cms-components.service';
 export * from './cms-features.service';
+export * from './cms-guards.service';
+export * from './cms-i18n.service';
 export * from './cms-routes.service';
