+++ projects/storefrontlib/shared/directives/lcp-context/index.ts
@@ -4,10 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from '../services/cms-lcp.service';
 export * from './lcp-context-directive.module';
 export * from './lcp-context.directive';
 export * from './lcp-context.model';
+export * from './lcp-context.token';
 export * from './lcp-presence-mapping.service';
 export * from './provide-lcp-context-directive.module';
 export * from './provide-lcp-context.directive';
